:bug: Fix ghost pickable items bug
This bug was introduced when migrating the codebase to C# from the
GDScript language. In Pickable.OnBodyEntered, the call to Teardown
was missing, which is what causes the fade out animation and the
freeing from memory. The Teardown call has been reintroduced into
the Pickable.OnBodyEntered method and works as intended.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>