:sparkles: Implement basic pause menu
The pause menu can be activated by pressing Start/Options on a
controller or by pressing the Escape key on a computer. The pause
menu allows the player to either resume execution of the current
game loop, restart the level, or go back to the main menu.

As of this moment, there is no mobile UI in place to trigger the
pause menu, but this should get taken care of in a future release.

^SPR-2 stage Review work Development 20m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>