:sparkles: Implement score setting and resets on Game Over screen
The Game Over screen will now properly display the score and time
remaining when the player reaches the end of the game. In Endless,
the time is not displayed, given that there is no time limint. For
cases where the player trips over a palette and loses the level,
the time is displayed along with the score.

When the game over screen is displayed, the previous scores are
recored from the game loop into GameState. When the player restarts
the level or goes to the main menu, those values are cleared.

^SPR-9 stage Review work Development 10m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>