:recycle: Change how signals are connected
In previous commits, custom signals that were connected to buttons
and other events in the game were given the string representation
of the callback method names. However, this could lead to a problem
where the name of the callback method is changed, and it also
presents the dreaded IDE0051 warning, indicating that the method in
question is never used, even though the game uses it as a callback.

To resolve this, the nameof function is called on each of the
callback methods as the last parameter in the Connect method. For
example:

Connect("body_entered", this, nameof(OnBodyEntered));

This effectively also allows for easy refactoring when needing to
change the name of the callback methods.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>