:recycle: Change how levels are parsed
This commit refactors the level data parsing code to utilize a
struct to host the data instead of a Godot Dictionary that needs to
be cast.

Additionally, the objects in the game now correctly point to the
recently relocated scripts. Finally, a bug where the game could not
find the queue_free method on Area2D has been fixed.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>