:sparkles: Add a starting timer to the game loop
The starting timer freezes the player from movement and is used to
let the player prepare for the level more easily. After three
seconds, the player is able to move freely and start the level.

Additionally, the scenes folder has been reorganized to better
contain some of the scenes within.

This is a step in making the game more playable and may mitigate the
issues outlined in SPR-6.

^SPR-6 stage Developer work Development 10m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>