:recycle: Add new extension methods
The GetCurrentState extension method has been added to get the
GameState root node without repeating the same assignment in the
codebase.

Additionally, the game will refuse to spawn hazards on the first
level to allow the player to get accustomed to playing the game
without avoiding obstacles first, another mitigation for SPR-6.

^SPR-6 work Development 15m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>