:children_crossing: Update tutorial experience
The HUD has been updated to show different tutorial texts on each
tutorial level. This allows the player to get familiar with the
objectives of the game, the controls, and what each powerup does
in the game.

The tutorial is broken up into the first three levels:
- Basic movement and collection of packages
- Using timepieces to increase the timer
- Being wary of hazards

I believe this transition will make the player feel much more
comfortable with the game. Additionally, the tutorial text now
matches the other UI elements, being consistent with the theme.

The game loop has also been updated to match the tutorial level
constraints.

^SPR-10 stage Review work Development 15m
^SPR-7 work Development 10m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>