:children_crossing: Make powerups more distinct
Powerups now emit their own light to make them easier to see on the
factory floor. Additionally, for package powerups, a new piece of
text displays that shows their value representation when a player
collides with it; i.e., '+1' for standard packages, and '+2' for
plus packages.

The collision radius of the package has also been increased to match
the light emission radius, which should make it easier for them to be
attained.

^SPR-7 work Development 10m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>