:children_crossing: Update timepiece algorithms
In previous iterations of the game, the timepiece modifier would add a
constant seven seconds to the timer without respect to anything else in
the game. This could potentially allow the player to rack up an absurd
amount of time to complete the level, which may not be desired from a
challenge standpoint.

The timepiece modifier has been adjusted to account for this. It uses a
form of exponential growth/decay to determine how much time to add on
to the timer. The equation is as follows:

t = packagesRemaining (t_lastpack - t_level) ^ growthrate

Where t_lastpack is the timestamp at which a package was last collected.
Growth rate is also controlled in the Godot editor as a hyperparameter.

Additionally, GameLoop has been refactored somewhat to make the file
smaller.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>