:children_crossing: Add a fadeout on success
In previous iterations of the game, the level success screen would
appear immediately when the player collects all of the packages.
However, this feels a little weird and unexpected for the player.

To resolve this issue, the level now fades out once the player reaches
the package requirements before showing the level success screen.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>