:recycle: Make GameMode a public property
The getter and setter methods for the CurrentGameMode in the GameState
class have been marked as obsolete, because the getters and setters
don't do much besides setting the respective property. As such, it makes
more sense to let the developer access the GameMode directly as a public
property: GameState.CurrentGameMode.

This does not affect the CurrentLevel, since there is additional logic
involved that necessitates the usage of getters and setters.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>