:recycle: Move tween fadeouts
A really common action a tween interpolates is a fadeout from fully
opaque to completely transparent. This code has been duplicated in a lot
of areas in the code because of how common this fadeout is.

To trim boilerplate code that has been duplicated, a new Tween extension
has been added that handles the interpolation correctly.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>